[Refactor] 탭별 메인제목 컴포넌트화
+ components 폴더 정리
